In `BP_Attachment->upload()`, only filter the WordPress upload_dir if the callback function is set.
This callback function can be passed as the second argument of `BP_Attachment->upload()`, or dynamically created if the specific `$base_dir` parameter was provided by the extending class.

About this second case, if the extending class forgets to set the `$base_dir` parameter or do it on purpose in order to use the WordPress default upload directory, we will not filter the WordPress upload_dir to avoid a warning.

This commit also adds a third argument to `BP_Attachment->upload()`, in case there is a need to pass a specific time (yyyy/mm) to set the WordPress default upload sub directory.

Fixes #6438

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9874 cdf35c40-ae34-48e0-9cc9-0c9da1808c22