Messages: Remove unused redirect in `bp_has_message_threads()`.
This change prevents a debug notice for themes with built-in BuddyPress support, by removing a tightly-scoped redirect attempt that only occurs after page output has already started. It also removes a superfluous `bp_moderate` capability check from the message notices screen function, as that function is only hooked in if the user has `bp_moderate` previously.

Fixes #6506 (in trunk, for 2.4)

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9946 cdf35c40-ae34-48e0-9cc9-0c9da1808c22