Activity: When calculating the activity comment depth, prevent an infinite loop when the parent activity item is not an activity update.
In #2768, we added the ability to calculate the activity comment depth.
The code at the time was meant to calculate the activity comment depth for
regular activity updates (`'activity_update'`).  However, it is possible
for other activity types to have activity comments attached to them.  For
example, since the introduction of comment syncing in 2.0, it is now
possible for activity comments to be attached to `'new_blog_post'` items.

When fetching activity comments attached to other activity types (as in the
case when using the "New Comments" activity dropdown filter), an infinite
loop can occur while traversing up the activity tree to calculate the
activity comment depth.

This commit fixes this by checking if the parent activity item is not an
activity update during activity comment depth calculation.  If the parent
activity is not an activity update, we have reached the top of the tree can
bail from the recursive depth check.

Fixes #6518.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9961 cdf35c40-ae34-48e0-9cc9-0c9da1808c22