Activity: fix @mentions in wp-admin post editor.
If the screen loads with the text editor visible, `bp_mentions` is attached to the text editor as expected, but `loadMentionsInTinyMCE` uses up its tries and gives up trying to attach the iframe, so when you switch to the visual editor, ` doesn't work.

This change removes the looped `loadMentionsInTinyMCE` approach and instead hooks into TinyMCE configuration to specify an init callback, where we then load `bp_mentions`.

Fixes #6487 (trunk). Props dcavins, imath, DJPaul.


git-svn-id: http://buddypress.svn.wordpress.org/trunk@10054 cdf35c40-ae34-48e0-9cc9-0c9da1808c22