Make sure the `xprofile_avatar_uploaded` action is fired once the avatar of the user is set.
In 2.3.0, we introduced a new UI to manage avatar. It now uses Ajax to upload and set the avatar. Unfortunately a confusion was made about the `xprofile_avatar_uploaded` action. It was firing immediately after the avatar was uploaded but before it has been set using the cropping tool.

This commit is:
- making sure the action is fired at the right time
- adding new parameters to the action to inform about the User ID and the way the avatar was set ("crop" or "camera")
- making sure an activity will be generated if the avatar was set within the wp-admin/extended-profile.

Props r-a-y, landwire

Fixes #6612 (trunk)

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10135 cdf35c40-ae34-48e0-9cc9-0c9da1808c22