Allow Administrators to disable/enable the Cover Image Feature from the BuddyPress settings (Administration screen)
- Create two new options for targeted components (xProfile & Groups)
- Generate two new setting fields into the components setting sections as soon as the feature is active. By default, the two new checkboxes are enabled.
- Introduce two new functions to check if the Administrator disallowed cover image uploads for Users (`bp_disable_cover_image_uploads()`) and Groups (`bp_disable_group_cover_image_uploads()`).
- Edit the `bp_is_active()` function to take in account the particularity of the xProfile component. See below for a detailled explanation.
- Introduce a new conditional tag to check if the current page is the profile "change-cover-image" screen.

The `bp_is_active()` function first checks into the "bp-active-components" database option to see if the component is active. In this option the xProfile component is keyed "xprofile". But the component global holding the extended BP_Component class is keyed "profile" and features are inside a property of this component global. So to be able to check if a feature attached to the xProfile component is active, we needed to edit this function changing the component key from "xprofile" to "profile". As a result, the xProfile filter to deactivate/activate the cover image is `bp_is_profile_cover_image_active`. For the Groups component, the filter is `bp_is_groups_cover_image_active`.

Props mercime, r-a-y

See #6570

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10152 cdf35c40-ae34-48e0-9cc9-0c9da1808c22