Avatar UI: make sure to display the full content of the image during the cropping step.
When a user uploads an image to set his avatar, we are saving a temporary version of it resized using the `bp_core_avatar_original_max_width()` dimension (by default 450px). This resized image will be output to the user so that he can edit it during the cropping step. Once the avatar is saved, we are deleting this temporary image.

Sometimes, the Avatar UI width is smaller than this resized version of the image and as a result the image is truncated on its right side (this mainly happens on mobile devices). To avoid this, the Avatar UI is now informing about its available width and if it is smaller than `bp_core_avatar_original_max_width()`. This available width will be used to generate the tempory image so that the full content of the image will be displayed to the user when cropping his avatar.

Fixes #6586

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10206 cdf35c40-ae34-48e0-9cc9-0c9da1808c22