URI: Allow root profiles to work when the 'Members' page is nested.
Previously, if the Members page was nested as a child of another page, the
root profiles feature - `define( 'BP_ENABLE_ROOT_PROFILES', true );` - did
not work.  This commit fixes this and includes a unit test.

Commit also adds a minor performance tweak when a root profile is detected.

Props r-a-y, boonebgorges.

Fixes #6477.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10212 cdf35c40-ae34-48e0-9cc9-0c9da1808c22