xprofile: fix PHP error on wp-admin Profile Fields screen
If a field ID is passed to the `BuddyPress_XProfile_Field` constructor,
that object is populated with that fields data. If the field ID is not
set (null), an empty profile field object would still be returned.

The change in r10200 to using `xprofile_get_field` does not (and should
not) return an empty profile field object if the field ID argument is
null, so this change adds appropriate handling for this situation.

Fixes #10200

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10226 cdf35c40-ae34-48e0-9cc9-0c9da1808c22