Make sure search terms are passed when rendering pagination.
r10110 altered the search query parameter for each BuddyPress component
(see #5087).  However, we also need to change the search query parameter in
our template loops so pagination will work when a search is in effect.

This commit alters the Members and Groups component template loops to use
the new search query parameter and alters the JS to find the correct search
term value.

See #6745 (trunk).

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10378 cdf35c40-ae34-48e0-9cc9-0c9da1808c22