Post Type Activities: allow post types to register comments tracking
When registering a post type, it is now possible to inform BuddyPress the comments about this post type need to be tracked into the activity stream. This can be achieved by making sure the post type supports "comments" and "buddypress-activity" and by adding extra parameters to the "bp_activity" argument of the `register_post_type()` function one or by using the `bp_activity_set_post_type_tracking_args()` function.

We are also introducing a new function to inform if an activity type is supporting a specific feature: `bp_activity_type_supports()`. For now this function is helping to check if:
- a activity supports a an activity comment
- an activity comment supports synchronization with a post type comment

Props shanebp, r-a-y.

See #6482

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10542 cdf35c40-ae34-48e0-9cc9-0c9da1808c22