Fix inaccurate timestamps caused by page caching.
If any template with a timestamp is statically cached, then the timestamps quickly become inaccurate.
The change adds livestamp.js and moment.js to dynamically update timestamps with the real relative time.

A consequence is that the labels surrounding the timestamps, notably in our widgets, have slightly changed.
We're going to keep an eye on this for the remainder of the 2.7 development cycle, and may make further adjustments as testing proves necessary.

Fixes #5757

Props r-a-y, imath, DJPaul

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11008 cdf35c40-ae34-48e0-9cc9-0c9da1808c22