Move composer.json to root and update Grunt.
Basic Composer support was added in v2.2, enabling distribution of BuddyPress via Packagist.
As we had no plans to use any developer-facing features of Composer, the config file was placed into the `/src/` subfolder (i.e. the root folder of the distribution version of BuddyPress) to simplify our release build process.

Moving the Composer config file into the root of the development version of the plugin, while copying it back into place for the distribution version, smooths the way towards enabling us to use Composer to install development-only libraries.

This may help us maintain a high-quality codebase. An example tickets where Composer is being considered for use as a development aid is #7228.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11036 cdf35c40-ae34-48e0-9cc9-0c9da1808c22