BuddyPress, Emails: set value of `bp-emails-unsubscribe-salt` option for new installs.
The initial value was set in `bp_get_default_options()` and worked because that function adds missing options to the database when it's ran.
For version updates, the new option was already being explictly added elsewhere.

However if a plugin is active and loads before the `plugins_loaded` hook, and calls add_option()` or `update_option()` (or multisite equivalents), a fatal error is caused by the `bp_core_clear_root_options_cache()` function.
This is because `wp_generate_password()` in used in `bp_get_default_options()` to set the `bp-emails-unsubscribe-salt` option, which is only available via `wp-includes/pluggable.php`. That file is only loaded by WordPress after plugins are loaded.

This change sets the default value in `bp_get_default_options()`, and then explicitly updates the value to the unsubscribe halts in the existing email installation function.

Fixes #7227
See #7254

Props r-a-y, DJPaul

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11093 cdf35c40-ae34-48e0-9cc9-0c9da1808c22