WP Admin Groups List: Add group type bulk change controls and filters.
Add a dropdown select to change the group type of many groups in
bulk. Also add a new column, “Group Type,” showing the group’s
assigned group type as a link that leads to a filtered view of the
groups list table, much like a tag link on the posts table works.

Props dcavins, djpaul.

Fixes #7175.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11139 cdf35c40-ae34-48e0-9cc9-0c9da1808c22