Fix incorrect `remove_action` and `remove_filter` usage for WordPress 4.7.
`remove_action` and `remove_filter` formally accepted a fourth parameter until WP 4.7; it's only with the introduction of the `WP_Hook` class that the function signature changed. Prior to WP 4.7, the fourth parameter was unused, so it does no harm to remove it and tidy things up.

Fixes #7318

Props ketuchetan

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11256 cdf35c40-ae34-48e0-9cc9-0c9da1808c22