Improve avatar uploader button click area for MS Edge Browsers
In ME Edge the Moxie shim click event is out of alignment with the visual button element causing it to appear that one can't click to open file browser.

Commit adjusts the styles to provide a moxie-shim set of rulesets in avatar.css to force the position to better overlap the button element.

This is a temp/perm fix until such time as the uploader group of elements can be re-factored to allow greater positioning control.

See #7360
Props angeljs, mercime, hnla, DJPaul, slaFFik

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11265 cdf35c40-ae34-48e0-9cc9-0c9da1808c22