When determining canonical URL, consider member type directories when 'page_on_front'.
`bp_get_canonical_url()` has some special logic for when
'page_on_front' is a BP component directory. This logic was introduced
before member and group type directories, so it did not account for
them.

The current changeset ensures that URLs of the format
example.com/members/type/foo do not redirect to example.com when
the Members directory is set to the home page.

The same problem doesn't arise for group type directories, because
the component/action stack works differently in the case of groups.

Fixes #7373.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11281 cdf35c40-ae34-48e0-9cc9-0c9da1808c22