Core: Redirect login links to their rightful spot if already authenticated.
Previously, if a logged-in user clicked on a login link with the
`'redirect_to'` parameter, the user would need to re-authenticate in order
to get redirected to the given URL.

This commit bypasses the re-authentication process for logged-in users,
which improves the user experience for those clicking on login links via
email.  `wp_safe_redirect()` is used to avoid redirecting to external
links that are not whitelisted by WordPress.

See #7349.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11610 cdf35c40-ae34-48e0-9cc9-0c9da1808c22