Remove logged-in check from `bp_activity_remove_all_user_data()`.
This function is called on the `delete_user` and `wpmu_delete_user` hooks.
During normal operation, these hooks only fire for users with the
`delete_users` cap. So the `is_user_logged_in()` check is not necessary.
It also interferes with the ability to delete users fully in an
automated fashion, as by script.

Introduced in [3753].

Fixes #7621.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11738 cdf35c40-ae34-48e0-9cc9-0c9da1808c22