Blogs: Disable activity commenting for blog items requiring manual approval.
For sites that have the "Comment must be manually approved" option enabled,
we are disabling activity commenting for these activity entries since there
is currently no way for us to replicate the WordPress pending comment
experience within the BuddyPress activity stream.

Users will still be able to comment on the original blog post, but will no
longer be able to comment within the BuddyPress activity stream.

Hat-tip aleon4 for reporting this issue.

Fixes #7592.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11788 cdf35c40-ae34-48e0-9cc9-0c9da1808c22