Change the display name WP/BP sync hook.
Instead of firing the display name synchronization routine on the general hook `xprofile_updated_profile`, which is run when the profile screen forms are submitted, use the lower level hook `xprofile_data_after_save` which is run when a profile field is updated. This catches more cases where we'd like the sync to run, like programatically updating the xprofile field.

See #7661.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11817 cdf35c40-ae34-48e0-9cc9-0c9da1808c22