Templates: move common assets from bp-legacy and bp-nouveau into a shared folder.
In the Nouveau and Legacy packs, these templates were identical. As they used for very specific parts of a BuddyPress theme,
and all to support features that were added since bp-legacy was created, we can utilise this new shared folder to avoid duplication.

The shared folder is added to the template hierarchy, with a lower priority than the WordPress theme, any WordPress parent theme, 
and the active template pack. Any site that has customised any of the moved templates will continue to load those version.

Fixes #7672

Props r-a-y for the initial patch, and hnla.


git-svn-id: http://buddypress.svn.wordpress.org/trunk@11835 cdf35c40-ae34-48e0-9cc9-0c9da1808c22