Templates: swap shared `_attachments/avatars/index.php` with Legacy version; fix path in wp-admin.
The Nouveau template was previously used, and included a call to a Nouveau-only function, which I missed in r11835.
The Nouveau pack now has a copy of its Nouveau template again.

Also changes a patch in `bp_attachments_get_template_part()` to support the changed location for this template asset.
This is used for the "Upload Avatar" screens in wp-admin Users > Edit.

Props boonebgorges, hnla.

Fixes #7679 and #7675



git-svn-id: http://buddypress.svn.wordpress.org/trunk@11839 cdf35c40-ae34-48e0-9cc9-0c9da1808c22