Unit Tests: Improve performance of multisite directory page test.
The unit test in question was introduced in #6226.

Basically, the test needed to match a specific WordPress post ID.
Previously, the test did this by creating posts in a loop(!) until it
matched the post ID.  However, there is a parameter in `wp_insert_post()`
that allows you to set the post ID from the get-go - `'import_id'`.

This commit switches to using the `'import_id'` parameter to create the
WordPress post and should speed up this test considerably.

Fixes #7709.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11880 cdf35c40-ae34-48e0-9cc9-0c9da1808c22