Activity: Move some functions from `bp-activity-actions.php` to `bp-activity-functions.php`
Some of the functions from `bp-activity-actions.php` are not related to
activity page actions, so these are being moved to the all-purpose
`bp-activity-functions.php` file (for better or worse!).

The intent is to remove and split up the `bp-activity-actions.php` and
`bp-activity-screens.php` files until these functions are actually
needed.

See #7218.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11883 cdf35c40-ae34-48e0-9cc9-0c9da1808c22