Activity: Conditionally load action and screen functions.
This commit conditionally loads action and screen function code for the
Activity component, utilizing the `'bp_late_include'` hook introduced in
r11884.

Previously, we loaded these functions at all times, which is unnecessary
when a user is not on an BuddyPress activity page.  Now, we only load this
code when needed.

A special case is required for PHPUnit as PHPUnit doesn't like conditional-
included files when running multiple unit tests.  So for PHPUnit, we
include all action and screen code at once, just like before.

See #7218.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11885 cdf35c40-ae34-48e0-9cc9-0c9da1808c22